+++ src/action.test.ts
@@ -36,6 +36,7 @@ type Inputs = {
   'update-only'?: string
   'comment-target'?: string
   'commit-sha'?: string
+  'delete-on-status'?: string
   preformatted?: string
   find?: string
   replace?: string
@@ -65,6 +66,7 @@ let getCommitCommentsResponse: Record<string, unknown>[] | undefined
 let postIssueCommentsResponse = {
   id: 42,
 }
+const deleteIssueCommentResponse = {}
 
 type MessagePayload = {
   comment_id?: number
@@ -119,6 +121,9 @@ const handlers = [
       return HttpResponse.json(postIssueCommentsResponse)
     },
   ),
+  http.delete(`https://api.github.com/repos/:repoUser/:repoName/issues/comments/:commentId`, () => {
+    return HttpResponse.json(deleteIssueCommentResponse)
+  }),
   http.get(`https://api.github.com/repos/:repoUser/:repoName/issues/:issueNumber/comments`, () => {
     return HttpResponse.json(getIssueCommentsResponse)
   }),
@@ -845,5 +850,42 @@ describe('commit comments', () => {
     // Delete old + create new = comment-updated (the re-created one)
     expect(core.setOutput).toHaveBeenCalledWith('comment-updated', 'true')
     expect(core.setOutput).toHaveBeenCalledWith('comment-id', 42)
+  })
+})
+
+describe('delete on status', () => {
+  it('can delete comment if status is matching', async () => {
+    inputs['delete-on-status'] = 'success'
+    inputs.status = 'success'
+    inputs.message = 'hello'
+
+    const body = `<!-- add-pr-comment:${inputs['message-id']} -->\n\n[ ] Hello\n[ ] World`
+
+    const commentId = 123
+
+    const replyBody = [
+      {
+        id: commentId,
+        body,
+      },
+    ]
+
+    getIssueCommentsResponse = replyBody
+
+    await run()
+
+    expect(core.setOutput).toHaveBeenCalledWith('comment-deleted', 'true')
+  })
+
+  it('does not delete comment if status is not matching', async () => {
+    inputs['delete-on-status'] = 'success'
+    inputs.status = 'failure'
+    inputs.message = 'hello'
+
+    getIssueCommentsResponse = []
+
+    await run()
+
+    expect(core.setOutput).toHaveBeenCalledWith('comment-created', 'true')
   })
 })