+++ src/types.ts
@@ -24,6 +24,7 @@ export interface Inputs {
   status: string
   owner: string
   updateOnly: boolean
+  deleteOnStatus?: string
 }
 
 export type CreateIssueCommentResponseData =